- added UntranslatedText and moved Text to own file
