+++ storage/Devices/DeviceImpl.cc
@@ -269,84 +269,84 @@ namespace storage
     Text
     Device::Impl::do_create_text(Tense tense) const
     {
-	return _("error: stub function called");
+	return UntranslatedText("error: stub do_create_text called");
     }
 
 
     void
     Device::Impl::do_create() const
     {
-	ST_THROW(LogicException("stub function called"));
+	ST_THROW(LogicException("stub do_create called"));
     }
 
 
     Text
     Device::Impl::do_delete_text(Tense tense) const
     {
-	return _("error: stub function called");
+	return UntranslatedText("error: stub do_delete_text called");
     }
 
 
     void
     Device::Impl::do_delete() const
     {
-	ST_THROW(LogicException("stub function called"));
+	ST_THROW(LogicException("stub do_delete called"));
     }
 
 
     Text
     Device::Impl::do_activate_text(Tense tense) const
     {
-	return _("error: stub function called");
+	return UntranslatedText("error: stub do_activate_text called");
     }
 
 
     void
     Device::Impl::do_activate() const
     {
-	ST_THROW(LogicException("stub function called"));
+	ST_THROW(LogicException("stub do_activate called"));
     }
 
 
     Text
     Device::Impl::do_deactivate_text(Tense tense) const
     {
-	return _("error: stub function called");
+	return UntranslatedText("error: stub do_deactivate_text called");
     }
 
 
     void
     Device::Impl::do_deactivate() const
     {
-	ST_THROW(LogicException("stub function called"));
+	ST_THROW(LogicException("stub do_deactivate called"));
     }
 
 
     Text
     Device::Impl::do_resize_text(ResizeMode resize_mode, const Device* lhs, Tense tense) const
     {
-	return _("error: stub function called");
+	return UntranslatedText("error: stub do_resize_text called");
     }
 
 
     void
     Device::Impl::do_resize(ResizeMode resize_mode) const
     {
-	ST_THROW(LogicException("stub function called"));
+	ST_THROW(LogicException("stub do_resize called"));
     }
 
 
     Text
     Device::Impl::do_reallot_text(ReallotMode reallot_mode, const Device* device, Tense tense) const
     {
-	return _("error: stub function called");
+	return UntranslatedText("error: stub do_reallot_text called");
     }
 
 
     void
     Device::Impl::do_reallot(ReallotMode reallot_mode, const Device* device) const
     {
-	ST_THROW(LogicException("stub function called"));
+	ST_THROW(LogicException("stub do_reallot called"));
     }
