+++ storage/Utils/AppUtil.cc
@@ -397,52 +397,6 @@ string afterLast(const string& s, const string& pat )
     }
 
 
-    void
-    Text::clear()
-    {
-	native.clear();
-	text.clear();
-    }
-
-
-    const Text&
-    Text::operator+=(const Text& a)
-    {
-	native += a.native;
-	text += a.text;
-	return *this;
-    }
-
-
-    Text
-    sformat(const Text& format, ...)
-    {
-	Text text;
-	va_list ap;
-
-	va_start(ap, format);
-	text.native = sformat(format.native, ap);
-	va_end(ap);
-
-	va_start(ap, format);
-	text.text = sformat(format.text, ap);
-	va_end(ap);
-
-	return text;
-    }
-
-
-    Text _(const char* msgid)
-    {
-	return Text(msgid, dgettext("libstorage-ng", msgid));
-    }
-
-    Text _(const char* msgid, const char* msgid_plural, unsigned long int n)
-    {
-	return Text(n == 1 ? msgid : msgid_plural, dngettext("libstorage-ng", msgid, msgid_plural, n));
-    }
-
-
     string
     hostname()
     {
